+++ src/main/java/it/blackhat/symposium/managers/QuestionManager.java
@@ -25,12 +25,24 @@ public interface QuestionManager {
     int insertQuestion(Question question) throws SQLException;
 
     /**
+     * Insert a tag in a question
+     *
+     * @param question question data
+     * @param tag      tag data
+     * @return the number of row updated
+     * @throws SQLException if it encounter an error
+     **/
+    int insertQuestionTag(Question question, Tag tag) throws SQLException;
+
+    /**
      * Delete a question
      *
      * @param questionId id of the question
      * @return the number of row updated
      * @throws SQLException if it encounter an error
      **/
+
+
     int deleteQuestion(int questionId) throws SQLException;
 
     /**
